more utility fixes

+++ PeregrineBot/Source/Utility.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "Common.h"
 
-float DistanceAir(const BWAPI::Position p1, const BWAPI::Position p2);
-float DistanceAir(const BWAPI::TilePosition start, const BWAPI::TilePosition end);
+double DistanceAir(const BWAPI::Position p1, const BWAPI::Position p2);
+double DistanceAir(const BWAPI::TilePosition start, const BWAPI::TilePosition end);
 BWAPI::Position GetPos(const BWAPI::TilePosition tp, const BWAPI::UnitType ut);
 BWAPI::Position GetBasePos(const BWAPI::TilePosition tp);
 double DistanceGround(const BWAPI::TilePosition start, const BWAPI::TilePosition end);
